+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* [CHANGE/BUGFIX] Disable compaction of overlapping blocks when Thanos sidecar is configured with object storage. #7747
+
 ## 0.84.0 / 2025-07-14
 
 * [FEATURE] Add `telegram` field to AlertManager CRD global configuration. #7631